refactor: flatten Config::Settings into Legion::LLM::Settings (v0.8.6)
Moves all default settings out of Legion::LLM::Config::Settings and into
Legion::LLM::Settings directly. Deletes lib/legion/llm/config/settings.rb
and the Config::Settings alias shim. legion/llm.rb and config.rb updated
to reference Settings.default. LegionIO service.rb already calls
Legion::LLM::Settings.default — no change needed there.
