+++ lib/legion/llm/version.rb
@@ -2,6 +2,6 @@
 
 module Legion
   module LLM
-    VERSION = '0.8.5'
+    VERSION = '0.8.6'
   end
 end
